refactor: separated cloud and edge viewcontroller
